requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
langchain>=0.2.0
openai>=1.30.0
